adding auto data collector ... however ...
(1) I should try Sanjay/Michael's simpler methods for needle detection and tracking. I'll still need to apply these on an arbitrary range of images, so this data collector will be useful anyway.
(2) We also need to change some of the ranges and height, we'll have to experiment with these. I think we also need to be careful about how extreme the angles can be, as it seems to be crashing frequently.

Still trying to get _something_ for March ...